requirements.txt
discord.py ~1.5.1
logging
datetime
validators
